editoast: move DetectorCache to infra_cache
