+++ editoast/src/infra_cache/mod.rs
@@ -31,6 +31,7 @@ use thiserror::Error;
 
 use crate::error::Result;
 use crate::infra_cache::object_cache::BufferStopCache;
+use crate::infra_cache::object_cache::DetectorCache;
 use crate::infra_cache::operation::CacheOperation;
 use crate::infra_cache::operation::RailjsonObject;
 use crate::map::BoundingBox;
@@ -732,11 +733,11 @@ pub mod tests {
     use diesel_async::scoped_futures::ScopedFutureExt;
     use editoast_schemas::infra::Waypoint;
 
-    use super::DetectorCache;
     use super::OperationalPointCache;
     use super::SignalCache;
     use super::TrackSectionCache;
     use crate::infra_cache::object_cache::BufferStopCache;
+    use crate::infra_cache::object_cache::DetectorCache;
     use crate::infra_cache::operation::create::tests::create_buffer_stop;
     use crate::infra_cache::operation::create::tests::create_detector;
     use crate::infra_cache::operation::create::tests::create_electrification;
