+++ editoast/src/infra_cache/object_cache.rs
@@ -1,3 +1,5 @@
 mod buffer_stop_cache;
+mod detector_cache;
 
 pub use buffer_stop_cache::BufferStopCache;
+pub use detector_cache::DetectorCache;
